#30082019
Добавлена продажа товара. Изменен алгоритм получения остатков с закупочной ценной. Добавлен партионный учет( пока что учитывается только продажа (без учета серийников))

+++ AagErp/ModelModul/Models/CountsProduct.cs
@@ -24,8 +24,8 @@
             }
         }
 
-        private double _count;
-        public double Count
+        private decimal _count;
+        public decimal Count
         {
             get => _count;
             set
