+++ AagErp/ModelModul/Models/Product.cs
@@ -122,8 +122,8 @@ namespace ModelModul.Models
             }
         }
 
-        private double _count;
-        public double Count
+        private decimal _count;
+        public decimal Count
         {
             get => _count;
             set
